src/chat-app/requirements.txt
chainlit ~= 2.0.6
langchain-openai ~= 0.2.1
langchain ~= 0.3.1
langchain-community ~= 0.3.1
python-dotenv ~= 1.0.1
azure-search-documents ~= 11.5.1
azure-identity ~= 1.18.0
langgraph ~= 0.2.32
pymssql~=2.3.2
azure-storage-blob
azure-ai-documentintelligence ~= 1.0.0
pymupdf ~= 1.25.2
unstructured
unstructured[pdf]
pillow
